+++ src/lib.rs
@@ -1,4 +1,3 @@
-#[cfg(feature = "client")]
 pub mod client;
 #[cfg(feature = "client")]
 pub mod frame;
@@ -10,8 +9,10 @@ pub mod server;
 mod test;
 pub mod types;
 
-use serde::{Deserialize, Serialize};
 use std::net::SocketAddr;
+
+use serde::{Deserialize, Serialize};
+#[cfg(any(feature = "client", feature = "server"))]
 use thiserror::Error;
 
 /// The error type for a handshake failure.