+++ src/server.rs
@@ -9,7 +9,7 @@ use oinq::{
 use quinn::Connection;
 use semver::{Version, VersionReq};
 
-use crate::{AgentInfo, HandshakeError};
+use crate::{client::RequestCode, AgentInfo, HandshakeError};
 
 /// Processes a handshake message and sends a response.
 ///
@@ -70,4 +70,29 @@ pub async fn handshake(
             version_req.to_string(),
         ))
     }
+}
+
+/// Sends a list of trusted domains to the client.
+///
+/// # Errors
+///
+/// Returns an error if serialization failed or communication with the client failed.
+pub async fn send_trusted_domain_list(conn: &Connection, list: &[String]) -> anyhow::Result<()> {
+    use anyhow::anyhow;
+    use bincode::Options;
+
+    let Ok(mut msg) = bincode::serialize::<u32>(&RequestCode::TrustedDomainList.into()) else {
+        unreachable!("serialization of u32 into memory buffer should not fail")
+    };
+    let ser = bincode::DefaultOptions::new();
+    msg.extend(ser.serialize(list)?);
+
+    let (mut send, mut recv) = conn.open_bi().await?;
+    frame::send_raw(&mut send, &msg).await?;
+
+    let mut response = vec![];
+    frame::recv_raw(&mut recv, &mut response).await?;
+    frame::recv::<Result<(), String>>(&mut recv, &mut response)
+        .await?
+        .map_err(|e| anyhow!(e))
 }